Extract the common part to calculate the layout for Flexbox as FlexboxHelper.
