+++ flexbox/src/main/java/com/google/android/flexbox/FlexboxLayout.java
@@ -21,10 +21,8 @@ import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.drawable.Drawable;
 import android.support.annotation.IntDef;
-import android.support.annotation.NonNull;
 import android.support.v4.view.ViewCompat;
 import android.util.AttributeSet;
-import android.util.SparseIntArray;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.LinearLayout;
@@ -73,7 +71,7 @@ import java.util.List;
  * <li>{@code layout_wrapBefore}</li>
  * </ul>
  */
-public class FlexboxLayout extends ViewGroup {
+public class FlexboxLayout extends ViewGroup implements FlexContainer {
 
     //TODO: Extracting the interfaces to independent classes is a breaking change. Update the
     //      document to notify that.
@@ -164,17 +162,7 @@ public class FlexboxLayout extends ViewGroup {
     /** The width of the {@link #mDividerDrawableVertical}. */
     private int mDividerVerticalWidth;
 
-    /**
-     * Holds reordered indices, which {@link LayoutParams#order} parameters are taken into account
-     */
-    private int[] mReorderedIndices;
-
-    /**
-     * Caches the {@link LayoutParams#order} attributes for children views.
-     * Key: the index of the view ({@link #mReorderedIndices} isn't taken into account)
-     * Value: the value for the order attribute
-     */
-    private SparseIntArray mOrderCache;
+    private FlexboxHelper mFlexboxHelper = new FlexboxHelper(this);
 
     private List<FlexLine> mFlexLines = new ArrayList<>();
 
@@ -241,8 +229,8 @@ public class FlexboxLayout extends ViewGroup {
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
         super.onMeasure(widthMeasureSpec, heightMeasureSpec);
-        if (isOrderChangedFromLastMeasurement()) {
-            mReorderedIndices = createReorderedIndices();
+        if (mFlexboxHelper.isOrderChangedFromLastMeasurement()) {
+            mFlexboxHelper.mReorderedIndices = mFlexboxHelper.createReorderedIndices();
         }
         if (mChildrenFrozen == null || mChildrenFrozen.length < getChildCount()) {
             mChildrenFrozen = new boolean[getChildCount()];
@@ -277,10 +265,10 @@ public class FlexboxLayout extends ViewGroup {
      * returns {@code null}.
      */
     public View getReorderedChildAt(int index) {
-        if (index < 0 || index >= mReorderedIndices.length) {
+        if (index < 0 || index >= mFlexboxHelper.mReorderedIndices.length) {
             return null;
         }
-        return getChildAt(mReorderedIndices[index]);
+        return getChildAt(mFlexboxHelper.mReorderedIndices[index]);
     }
 
     @Override
@@ -289,120 +277,11 @@ public class FlexboxLayout extends ViewGroup {
         // ViewGroup since otherwise reordered indices won't be in effect before the
         // FlexboxLayout's onMeasure is called.
         // Because requestLayout is requested in the super.addView method.
-        mReorderedIndices = createReorderedIndices(child, index, params);
+        mFlexboxHelper.mReorderedIndices =
+                mFlexboxHelper.createReorderedIndices(child, index, params);
         super.addView(child, index, params);
     }
 
-    /**
-     * Create an array, which indicates the reordered indices that {@link LayoutParams#order}
-     * attributes are taken into account. This method takes a View before that is added as the
-     * parent ViewGroup's children.
-     *
-     * @param viewBeforeAdded          the View instance before added to the array of children
-     *                                 Views of the parent ViewGroup
-     * @param indexForViewBeforeAdded  the index for the View before added to the array of the
-     *                                 parent ViewGroup
-     * @param paramsForViewBeforeAdded the layout parameters for the View before added to the array
-     *                                 of the parent ViewGroup
-     * @return an array which have the reordered indices
-     */
-    private int[] createReorderedIndices(View viewBeforeAdded, int indexForViewBeforeAdded,
-            ViewGroup.LayoutParams paramsForViewBeforeAdded) {
-        int childCount = getChildCount();
-        List<Order> orders = createOrders(childCount);
-        Order orderForViewToBeAdded = new Order();
-        if (viewBeforeAdded != null
-                && paramsForViewBeforeAdded instanceof FlexboxLayout.LayoutParams) {
-            orderForViewToBeAdded.order = ((LayoutParams) paramsForViewBeforeAdded).order;
-        } else {
-            orderForViewToBeAdded.order = LayoutParams.ORDER_DEFAULT;
-        }
-
-        if (indexForViewBeforeAdded == -1 || indexForViewBeforeAdded == childCount) {
-            orderForViewToBeAdded.index = childCount;
-        } else if (indexForViewBeforeAdded < getChildCount()) {
-            orderForViewToBeAdded.index = indexForViewBeforeAdded;
-            for (int i = indexForViewBeforeAdded; i < childCount; i++) {
-                orders.get(i).index++;
-            }
-        } else {
-            // This path is not expected since OutOfBoundException will be thrown in the ViewGroup
-            // But setting the index for fail-safe
-            orderForViewToBeAdded.index = childCount;
-        }
-        orders.add(orderForViewToBeAdded);
-
-        return sortOrdersIntoReorderedIndices(childCount + 1, orders);
-    }
-
-    /**
-     * Create an array, which indicates the reordered indices that {@link LayoutParams#order}
-     * attributes are taken into account.
-     *
-     * @return @return an array which have the reordered indices
-     */
-    private int[] createReorderedIndices() {
-        int childCount = getChildCount();
-        List<Order> orders = createOrders(childCount);
-        return sortOrdersIntoReorderedIndices(childCount, orders);
-    }
-
-    private int[] sortOrdersIntoReorderedIndices(int childCount, List<Order> orders) {
-        Collections.sort(orders);
-        if (mOrderCache == null) {
-            mOrderCache = new SparseIntArray(childCount);
-        }
-        mOrderCache.clear();
-        int[] reorderedIndices = new int[childCount];
-        int i = 0;
-        for (Order order : orders) {
-            reorderedIndices[i] = order.index;
-            mOrderCache.append(i, order.order);
-            i++;
-        }
-        return reorderedIndices;
-    }
-
-    @NonNull
-    private List<Order> createOrders(int childCount) {
-        List<Order> orders = new ArrayList<>();
-        for (int i = 0; i < childCount; i++) {
-            View child = getChildAt(i);
-            LayoutParams params = (LayoutParams) child.getLayoutParams();
-            Order order = new Order();
-            order.order = params.order;
-            order.index = i;
-            orders.add(order);
-        }
-        return orders;
-    }
-
-    /**
-     * Returns if any of the children's {@link LayoutParams#order} attributes are changed
-     * from the last measurement.
-     *
-     * @return {@code true} if changed from the last measurement, {@code false} otherwise.
-     */
-    private boolean isOrderChangedFromLastMeasurement() {
-        int childCount = getChildCount();
-        if (mOrderCache == null) {
-            mOrderCache = new SparseIntArray(childCount);
-        }
-        if (mOrderCache.size() != childCount) {
-            return true;
-        }
-        for (int i = 0; i < childCount; i++) {
-            View view = getChildAt(i);
-            if (view == null) {
-                continue;
-            }
-            LayoutParams lp = (LayoutParams) view.getLayoutParams();
-            if (lp.order != mOrderCache.get(i)) {
-                return true;
-            }
-        }
-        return false;
-    }
 
     /**
      * Sub method for {@link #onMeasure(int, int)}, when the main axis direction is horizontal
@@ -2454,7 +2333,7 @@ public class FlexboxLayout extends ViewGroup {
      */
     public static class LayoutParams extends ViewGroup.MarginLayoutParams {
 
-        private static final int ORDER_DEFAULT = 1;
+        static final int ORDER_DEFAULT = 1;
 
         private static final float FLEX_GROW_DEFAULT = 0f;
 
@@ -2620,35 +2499,6 @@ public class FlexboxLayout extends ViewGroup {
 
         public LayoutParams(int width, int height) {
             super(new ViewGroup.LayoutParams(width, height));
-        }
-    }
-
-    /**
-     * A class that is used for calculating the view order which view's indices and order
-     * properties from Flexbox are taken into account.
-     */
-    private static class Order implements Comparable<Order> {
-
-        /** {@link View}'s index */
-        int index;
-
-        /** order property in the Flexbox */
-        int order;
-
-        @Override
-        public int compareTo(@NonNull Order another) {
-            if (order != another.order) {
-                return order - another.order;
-            }
-            return index - another.index;
-        }
-
-        @Override
-        public String toString() {
-            return "Order{" +
-                    "order=" + order +
-                    ", index=" + index +
-                    '}';
         }
     }
 }
